# vectorize `find_first_of` for long needle 🪡