In benchmark results only the last line changed:

before
```
-----------------------------------------------------------------
Benchmark                       Time             CPU   Iterations
-----------------------------------------------------------------
bm<uint8_t, 2, 3>            2.70 ns         2.35 ns    597333333
bm<uint16_t, 2, 3>           2.71 ns         2.13 ns    689230769
bm<uint8_t, 7, 4>            14.1 ns         10.9 ns    112000000
bm<uint16_t, 7, 4>           12.4 ns         10.6 ns    100000000
bm<uint8_t, 9, 3>            9.77 ns         7.70 ns    190638298
bm<uint16_t, 9, 3>           10.4 ns         8.06 ns    162909091
bm<uint8_t, 22, 5>           10.1 ns         8.02 ns    208372093
bm<uint16_t, 22, 5>          11.0 ns         8.64 ns    151864407
bm<uint8_t, 3056, 7>          273 ns          233 ns      5973333
bm<uint16_t, 3056, 7>         561 ns          425 ns      3089655
bm<uint8_t, 1011, 11>        95.5 ns         79.2 ns     17568627
bm<uint16_t, 1011, 11>       2872 ns         2448 ns       497778
```
after
```
-----------------------------------------------------------------
Benchmark                       Time             CPU   Iterations
-----------------------------------------------------------------
bm<uint8_t, 2, 3>            2.71 ns         1.80 ns    746666667
bm<uint16_t, 2, 3>           2.71 ns         1.55 ns   1000000000
bm<uint8_t, 7, 4>            14.1 ns         7.47 ns    175686275
bm<uint16_t, 7, 4>           12.3 ns         5.82 ns    190638298
bm<uint8_t, 9, 3>            9.78 ns         6.75 ns    208372093
bm<uint16_t, 9, 3>           10.7 ns         7.54 ns    190638298
bm<uint8_t, 22, 5>           10.3 ns         7.21 ns    169056604
bm<uint16_t, 22, 5>          11.1 ns         8.63 ns    175686275
bm<uint8_t, 3056, 7>          276 ns          205 ns     11200000
bm<uint16_t, 3056, 7>         552 ns          432 ns      2890323
bm<uint8_t, 1011, 11>        94.4 ns         76.2 ns     19478261
bm<uint16_t, 1011, 11>        481 ns          365 ns      3895652
```