@AlexGuteniev 2024-04-03
> There are real test failures here.

There were multiple errors. 

After fixing all of them, I thought it would be better to preserve the <= 16 version, as the generic becomes noticeable slower.