@StephanTLavavej 2024-04-09
Thanks for improving your vectorized implementation even further! :rocket: :rocket: :rocket: